pyproject.toml
[tool.poetry]
name = "diet-vision"
version = "0.1.0"
description = "Diet Vision"
authors = ["Euijae Kim <ek3955@nyu.edu>"]
readme = "README.md"
packages = [{include = "apps"}]

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.7.0.72"
numpy = "^1.24.3"
torch = "^2.0.1"
segment-anything = "^1.0"
jupyter-bbox-widget = "^0.5.0"
roboflow = "^1.0.9"
dataclasses-json = "^0.5.8"
supervision = "^0.10.0"
datasets = "^2.13.0"
transformers = "^4.30.2"
pillow = "^9.5.0"
matplotlib = "^3.7.2"
gradio = "^3.36.1"
fastapi = "^0.100.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
